chore: Update check_price function in scraper/check_price.py
This commit adds an additional parameter `access_token` to the `check_price` function in `scraper/check_price.py`. The `access_token` is used to send notifications for the maximum and minimum values of ticker data. This change improves the functionality of the function by allowing it to send notifications using the provided access token.